Rename declares. Deprecate declare-cairo0.
