+++ protostar/commands/__init__.py
@@ -1,6 +1,6 @@
 from .cairo1_commands import (
     BuildCommand,
-    DeclareCairo1Command,
+    DeclareCommand,
     TestCommand,
     InitCommand,
 )
@@ -11,7 +11,7 @@ from .legacy_commands import (
 )
 from .calculate_account_address_command import CalculateAccountAddressCommand
 from .call import CallCommand
-from .declare import DeclareCommand
+from .legacy_commands.declare_cairo0 import DeclareCairo0Command
 from .deploy_account_command import DeployAccountCommand
 from .deploy_command import DeployCommand
 from .format_command import FormatCommand
