+++ protostar/commands/cairo1_commands/__init__.py
@@ -1,4 +1,4 @@
 from .build_command import BuildCommand
 from .init_command import InitCommand
 from .test_command import TestCommand
-from .declare_cairo1_command import DeclareCairo1Command
+from .declare_command import DeclareCommand
