--- protostar/commands/declare/__init__.py
@@ -1,1 +1,0 @@
-from .declare_command import DeclareCommand
