+++ protostar/commands/legacy_commands/declare_cairo0/__init__.py
@@ -1,0 +1,1 @@
+from .declare_cairo0_command import DeclareCairo0Command
