+++ tests/integration/_conftest/protostar_fixture_factory.py
@@ -19,8 +19,8 @@ from protostar.commands import (
     InitCairo0Command,
     InvokeCommand,
     MulticallCommand,
-    DeclareCairo1Command,
     TestCairo0Command,
+    DeclareCairo0Command,
 )
 from protostar.commands.deploy_account_command import DeployAccountCommand
 from protostar.commands.deploy_command import DeployCommand
@@ -147,11 +147,11 @@ def create_protostar_fixture(
         project_root_path=project_root_path,
         messenger_factory=messenger_factory,
     )
-    declare_command = DeclareCommand(
+    declare_cairo0_command = DeclareCairo0Command(
         gateway_facade_factory=gateway_facade_factory,
         messenger_factory=messenger_factory,
     )
-    declare_cairo1_command = DeclareCairo1Command(
+    declare_command = DeclareCommand(
         contract_path_resolver=contract_path_resolver,
         gateway_facade_factory=gateway_facade_factory,
         messenger_factory=messenger_factory,
@@ -218,7 +218,7 @@ def create_protostar_fixture(
         build_cairo0_command=build_cairo0_command,
         format_command=format_command,
         declare_command=declare_command,
-        declare_cairo1_command=declare_cairo1_command,
+        declare_cairo0_command=declare_cairo0_command,
         deploy_command=deploy_command,
         test_command=test_command,
         test_cairo0_command=test_cairo0_command,
