+++ tests/integration/commands/declare_command_test.py
@@ -1,24 +1,17 @@
+from argparse import Namespace
 from pathlib import Path
+from typing import Any
+from unittest.mock import MagicMock
 
 import pytest
-from starknet_py.net.gateway_client import GatewayClient
 from starknet_py.net.models import StarknetChainId
 
-from protostar.cli.signable_command_util import PRIVATE_KEY_ENV_VAR_NAME
-from protostar.starknet import Address
-from protostar.starknet_gateway import FeeExceededMaxFeeException
+from protostar.cli import MessengerFactory
+from protostar.commands import DeclareCommand
+from protostar.contract_path_resolver import ContractPathResolver
+from protostar.io import log_color_provider
+from protostar.starknet_gateway import GatewayFacadeFactory
 from tests.conftest import DevnetAccount, SetPrivateKeyEnvVarFixture
-from tests.data.contracts import CONTRACT_WITH_CONSTRUCTOR
-from tests.integration.conftest import CreateProtostarProjectFixture
-from tests.integration._conftest import ProtostarFixture
-
-
-@pytest.fixture(name="protostar")
-def protostar_fixture(create_protostar_project: CreateProtostarProjectFixture):
-    with create_protostar_project() as protostar_project:
-        protostar_project.create_files({"./src/main.cairo": CONTRACT_WITH_CONSTRUCTOR})
-        protostar_project.protostar.build_cairo0_sync()
-        yield protostar_project.protostar
 
 
 @pytest.fixture(name="compiled_contract_path")
@@ -26,69 +19,50 @@ def compiled_contract_path_fixture() -> Path:
     return Path("./build/main.json")
 
 
-async def test_declaring_contract(
-    protostar: ProtostarFixture,
+@pytest.fixture(name="mocked_contract_path_resolver")
+def mocked_contract_path_resolver_fixture(datadir: Path) -> ContractPathResolver:
+    class MockedContractPathResolver(ContractPathResolver):
+        def __init__(self):
+            super().__init__(MagicMock(), MagicMock())
+
+        def contract_path_from_contract_name(self, *args: Any, **kwargs: Any) -> Path:
+            # pylint: disable=unused-argument
+            return datadir
+
+    return MockedContractPathResolver()
+
+
+async def test_declaring_cairo1_contract(
     devnet_gateway_url: str,
     devnet_account: DevnetAccount,
-    compiled_contract_path: Path,
     set_private_key_env_var: SetPrivateKeyEnvVarFixture,
+    mocked_contract_path_resolver: ContractPathResolver,
 ):
-    with set_private_key_env_var(devnet_account.private_key):
-        response = await protostar.declare(
-            chain_id=StarknetChainId.TESTNET,
-            account_address=devnet_account.address,
-            contract=compiled_contract_path,
-            gateway_url=devnet_gateway_url,
-            max_fee="auto",
-        )
-
-        assert response.class_hash is not None
-
-
-async def test_max_fee_is_respected(
-    protostar: ProtostarFixture,
-    devnet_gateway_url: str,
-    devnet_account: DevnetAccount,
-    compiled_contract_path: Path,
-    set_private_key_env_var: SetPrivateKeyEnvVarFixture,
-):
-    with set_private_key_env_var(devnet_account.private_key):
-        with pytest.raises(FeeExceededMaxFeeException):
-            await protostar.declare(
-                chain_id=StarknetChainId.TESTNET,
-                account_address=devnet_account.address,
-                contract=compiled_contract_path,
-                gateway_url=devnet_gateway_url,
-                max_fee=1,
-                wait_for_acceptance=True,
-            )
-
-
-@pytest.mark.xfail(
-    reason="This test is going to fail since sending signed deploy txs is supported only for devnet, and now it's "
-    "disabled "
-)
-async def test_deploying_contract_with_signing(
-    devnet_gateway_url: str,
-    protostar: ProtostarFixture,
-    compiled_contract_path: Path,
-    monkeypatch: pytest.MonkeyPatch,
-):
-    monkeypatch.setenv(PRIVATE_KEY_ENV_VAR_NAME, "123")
-
-    response = await protostar.declare(
-        chain_id=StarknetChainId.TESTNET,
-        account_address=Address.from_user_input("123"),
-        contract=compiled_contract_path,
-        gateway_url=devnet_gateway_url,
-        wait_for_acceptance=True,
+    declare = DeclareCommand(
+        gateway_facade_factory=GatewayFacadeFactory(Path("")),
+        messenger_factory=MessengerFactory(
+            log_color_provider=log_color_provider,
+            activity_indicator=MagicMock(),
+        ),
+        contract_path_resolver=mocked_contract_path_resolver,
     )
 
-    assert response.class_hash is not None
+    args = Namespace(
+        chain_id=StarknetChainId.TESTNET,
+        account_address=devnet_account.address,
+        contract="minimal",
+        gateway_url=devnet_gateway_url,
+        max_fee=int(1e16),
+        wait_for_acceptance=False,
+        json=False,
+        signer_class=None,
+        private_key_path=None,
+        network=None,
+        token=None,
+        block_explorer=None,
+    )
 
-    gateway_client = GatewayClient(devnet_gateway_url)
-    transaction = await gateway_client.get_transaction(response.transaction_hash)
-    assert transaction.signature == [
-        "3459263272550625393812584460277149848351409720716906360199187355059506361232",
-        "1830633873577487268914590325347030490499699872026627682646826922183589844384",
-    ]
+    with set_private_key_env_var(devnet_account.private_key):
+        response = await declare.run(args)
+
+        assert response.class_hash is not None
