+++ tests/integration/commands/declare_command_test/cairo_project.toml
