+++ tests/integration/commands/declare_command_test/lib.cairo
