+++ tests/integration/commands/multicall_command_test.py
@@ -32,7 +32,7 @@ async def standard_contract_class_hash_fixture(
 ):
     account = devnet.get_predeployed_accounts()[0]
     with set_private_key_env_var(account.private_key):
-        declare_result = await protostar.declare(
+        declare_result = await protostar.declare_cairo0(
             contract=Path() / "build" / "main.json",
             account_address=account.address,
             chain_id=StarknetChainId.TESTNET,
